Add files via upload
Stock market is one of the major fields that investors are dedicated to, thus stock market price trend prediction is always a hot topic  from both financial and technical domains. In this project, my objective is to build a model with lower RMSE (Root Mean Square Error) value and best fitted graph to predict future price trend with lower risk of loss and good results.